0.6
**********************************************************************
**********************************************************************
Vimerito 2 v.0.6

Versiondetails:
**********************************************************************
1. Modules included. Modules are application parts which are completly
isolated. Controllers outside of a module can't use the moduls
classes. For modules it is possible to use external classes.
The buildup of a module is completly like an application, but it
must be located inside of an application.

2. The system-class VLang included. With this class you can controll
various languages. For this you can use the database or an
INI-File.

3. Now you can call the VJavaScript::__construct() with a file,
e.g.:
$js = new VJavaScript("file:path/to/the/javascript.js");

or with source:
$js = new VJavaScript("script: document.body=''");

4. The activ-recorder can return JSON, now.

5. The big VEvent-Bug removed. Events now called.

6. VRouter, Vimerito, VRequest, VLayout adjusted for modules.

7. VRouter::route rewritten.

8. Vimerito::__autoload rewritten. Bugs removed.